Add files via upload
program that was used during one of my class challenges. It will log into an FTP server and read the file permissions of files in a certain given directory and translate those permissions into a series of bits that will then be translated into 7- and 10-bit ASCII representations. For obvious reasons I have removed the FTP servers IP address so you will have to provide one.